Prepare for rearchitecture: Remove source from tip.
Remove source from tip in preparation for rollout of major re-design.

Due to the large number of changes, including refactoring of all the
major class names and their inheritance structure, it is not worth
trying to keep change history on the source files.  No attempt at
backwards compatibility is being made for the new release.

Hello alpha.