Outfit: Remove body part layer and status settings
Remove body part layer and status settings.  Having these settings as
part of the mandatory API confuses matters and limits flexibility.
Having these features in the Outfit API implies that all body parts are
expected to always have the same layer and status.  This is not the
case.  Also, having the standard outfit apply the settings at
initialization is inappropriate since it limits the flexibility of
outfit designs.