Accessory Redesign
Redesign the Accessory and its associated classes to improve ease-of-use
and flexibility.  Primary changes:

Remove 'unmounting' status.  All accessories are required to immeidately
unmount.  Allowing an unmount to have a duration increases complexity in
accessory managers, such as the body, without significant benefits.(A
manager should not be forced to wait to store its accessories.)

Refactor the 'not mounted' status to 'unmanaged'.  This more clearly
describes the state of the accessory.

Remove the 'invalid' status and add 'destroyed'.  This status is no
longer required.  (See below changes.)

Add a Destroy(DestroyType) method.  It will implement a standard method
for destroying an accessory and communicating its detruction to
observers and support components. Fold bake features into the destroy
feature as a destroy type.

Formalize ownership.  The accessory will manage all ownership changes
and an owner object will be required for all transitions to a managed
state, such as 'mounted' and 'stored'.  This change will make it easier
for owners and potential owners to track state change.  E.g. An owner
that is observing an accessory can easily tell when a re-mount includes
a transfer in ownership.

Remove the OnStatusChange() event and add an OnStateChange() event which
is sent whenver any of the following changes:  accessory status, mount
location, owner.  This combination of state is more valuable than just
the status.

Refactor Unmount() to Release() and remove Unstore().  The new design
will support transitions to/from any status.  Having two methods that
are both responsible for transitioning to the same status is confusing.