Migrate from AccesssoryMounter to IAccessoryMounter
Migrate from mandating that an accessory mounter be a MonoBehaviour to
supporting mounters of any type.

When serialization is required, an IAccessoryMounter of by
UnityEngine.Object will still be required.  But otherwise, any object
can implement mounter features.   This mostly applies to priority
mounter arguments.

Benefits:

Supports the use of project assets as mounters.  (More efficient the
having every instance of a prefab having its own group of mounters.)
Supports the use of simple and efficient priority mounters.

Drawback:  Probably won't be able to support serialization of
in-progress mounters.  This is not a big issue since there are no good
options for supporting that feature, even using the old component model.