Upgraded behavior of body when it loses its outfit.
Removed the concept of a body 'soft reset' and replaced it with a
'forced release'.   The forced release process is better at reducing the
negative side effects of a body unexpectedly losing its outfit.   It can
also be manually initiated by the end user to free the outfit from its
body ,in its current state, for operations such as baking.